Moved styles.ts file from pages/admin/transactions into features/transactions. Installed sharp package in order to optimize pictures. Removed api folder
