--- api/.editorconfig
@@ -1,2 +1,0 @@
-[*]
-ij_typescript_use_double_quotes = false
