--- api/public/uploads/.gitignore
@@ -1,1 +1,0 @@
-/avatars/*
