--- api/src/config.ts
@@ -1,5 +1,0 @@
-const config = {
-  db: 'mongodb://localhost/strategy-school',
-};
-
-export default config;
