--- api/src/users/create-users.dto.ts
@@ -1,7 +1,0 @@
-export class CreateUsersDto {
-  email: string;
-  password: string;
-  firstName: string;
-  lastName: string;
-  phoneNumber: string;
-}
