+++ frontend/package.json
@@ -36,6 +36,7 @@
     "react-redux": "^8.0.5",
     "react-share": "^4.4.1",
     "react-youtube": "^10.1.0",
+    "sharp": "^0.32.1",
     "typescript": "5.0.3"
   },
   "devDependencies": {
